% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\name{complete_covid}
\alias{complete_covid}
\title{Joins open data with the descriptions}
\usage{
complete_covid(input, descriptions)
}
\description{
Joins open data with the descriptions
}
\keyword{cases}
\keyword{covid19,}
\keyword{daily}
\keyword{mexico,}
