% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\name{get_cases_meta}
\alias{get_cases_meta}
\title{Get the meta for specific date, datset & subset}
\usage{
get_cases_meta(date = "today", dataset = "open", subset = "positive")
}
\description{
Get the meta for specific date, datset & subset
}
\keyword{cases}
\keyword{covid19,}
\keyword{daily}
\keyword{mexico,}
