% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_cities}
\alias{get_cities}
\title{Get the cities table}
\usage{
get_cities()
}
\description{
Get the cities table
}
\keyword{cities,}
\keyword{coordenadas,}
\keyword{coordinates}
\keyword{geographical}
\keyword{mexico,}
\keyword{municipios,}
\keyword{poblacion,}
\keyword{population,}
