% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\name{get_covid_cases}
\alias{get_covid_cases}
\title{Get the table of daily cases for COVID}
\usage{
get_covid_cases(date = "today", dataset = "open", subset = "positive")
}
\arguments{
\item{date}{Needs to be provided in ISO format YYYY-MM-DD.
Note that you might need to add
Sys.setenv(TZ = "America/Mexico_City") if you are in a
different timezone}

\item{dataset}{could be open, or "ctd", "ctd" stands for
"Comunicado Tecnico Diario" (Daily Technical Release).
The open data only exists after April 13th, 2020.}

\item{subset}{if dataset ctd is preferred, you need to select
a subset (positive or suspected), otherwise this argument
is ignored. In the future they may be condensed into a
single dataset}
}
\description{
Get the table of daily cases for COVID
}
\keyword{cases}
\keyword{covid19,}
\keyword{daily}
\keyword{mexico,}
