% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\name{get_covid_cases_meta}
\alias{get_covid_cases_meta}
\title{Get the meta for specific date, datset & subset}
\usage{
get_covid_cases_meta(date = "latest", dataset = "open", subset = "positive")
}
\arguments{
\item{date}{Needs to be provided in ISO format YYYY-MM-DD.
Note that you might need to add
Sys.setenv(TZ = "America/Mexico_City") if you are in a
different timezone. You can specify "today" but it is
better to leave it to the default "latest"}

\item{dataset}{could be open, or "ctd", "ctd" stands for
"Comunicado Tecnico Diario" (Daily Technical Release).
The open data only exists after April 13th, 2020.}

\item{subset}{if dataset ctd is preferred, you need to select
a subset (positive or suspected), otherwise this argument
is ignored. In the future they may be condensed into a
single dataset}
}
\description{
Get the meta for specific date, datset & subset
}
\keyword{cases}
\keyword{covid19,}
\keyword{daily}
\keyword{mexico,}
