% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\name{get_covid_descriptions}
\alias{get_covid_descriptions}
\title{Gets the open data descriptions}
\usage{
get_covid_descriptions()
}
\description{
Gets the open data descriptions
}
\keyword{cases}
\keyword{covid19,}
\keyword{daily}
\keyword{mexico,}
