% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\name{get_covid_meta}
\alias{get_covid_meta}
\title{Get the table of daily cases for COVID}
\usage{
get_covid_meta()
}
\description{
Get the table of daily cases for COVID
}
\keyword{cases}
\keyword{covid19,}
\keyword{daily}
\keyword{mexico,}
