% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\name{get_covid_timeseries}
\alias{get_covid_timeseries}
\title{Get time series for COVID-19 Cases in MX}
\usage{
get_covid_timeseries()
}
\description{
Get time series for COVID-19 Cases in MX
}
\keyword{covid19,}
\keyword{mexico,}
\keyword{timeseries}
