% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_states}
\alias{get_states}
\title{Get the states table}
\usage{
get_states()
}
\description{
Get the states table
}
\keyword{coordenadas,}
\keyword{coordinates}
\keyword{estados,}
\keyword{geographical}
\keyword{mexico,}
\keyword{poblacion,}
\keyword{population,}
\keyword{states,}
