% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\name{rename_old_covid_columns}
\alias{rename_old_covid_columns}
\title{Fixes structure prior to April 16}
\usage{
rename_old_covid_columns(df, warnings = FALSE)
}
\description{
Fixes structure prior to April 16
}
