% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries_all}
\alias{timeseries_all}
\title{Get time series for COVID-19 Cases in MX}
\usage{
timeseries_all()
}
\description{
Get time series for COVID-19 Cases in MX
}
\keyword{covid19,}
\keyword{mexico,}
\keyword{timeseries}
