% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries_totals}
\alias{timeseries_totals}
\title{Get time series for COVID-19 Cases in MX aggregated by date}
\usage{
timeseries_totals()
}
\description{
Get time series for COVID-19 Cases in MX aggregated by date
}
\keyword{covid19,}
\keyword{mexico,}
\keyword{timeseries}
