% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\name{validate_covid_structure}
\alias{validate_covid_structure}
\title{Validates structure of dataset}
\usage{
validate_covid_structure(df, stop = FALSE)
}
\description{
Validates structure of dataset
}
\keyword{cases}
\keyword{covid19,}
\keyword{daily}
\keyword{mexico,}
