% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{daily_cases}
\alias{daily_cases}
\title{Get time series for COVID-19 Cases in MX}
\usage{
daily_cases(date, type = "positive")
}
\arguments{
\item{date}{Needs to be provided in ISO format YYYY-MM-DD}

\item{type}{could be positive, suspected, positivos, sospechosos}
}
\description{
Get time series for COVID-19 Cases in MX
}
\keyword{cases}
\keyword{covid19,}
\keyword{daily}
\keyword{mexico,}
